SearchAsideBar component moved from SearchPage
